set up grouped vulns
